refactor(tone-analyzer): remove compatibility layer for tone analyzer
BREAKING CHANGE: Deprecated methods in Tone Analyzer are no longer available. Changed parameter names are no longer interally corrected.

To migrate your code, use the methods and parameters currently available with the service as documented here: https://www.ibm.com/watson/developercloud/tone-analyzer/api/v3/node.html?node
